refactor!: 重构 utils 模块
- 新增工具模块集，并移除原来单个 `util` 工具组件
- 新增 `layui/utils` 子路径导出，以支持外部按需引入工具模块
- 移除 `util.escape`、 `util.unescape` 方法
