--- src/css/modules/util.css
@@ -1,30 +1,0 @@
-/** fixbar **/
-.layui-fixbar {
-  position: fixed;
-  right: 16px;
-  bottom: 16px;
-  z-index: 999999;
-}
-.layui-fixbar li {
-  width: 50px;
-  height: 50px;
-  line-height: 50px;
-  margin-bottom: 1px;
-  text-align: center;
-  cursor: pointer;
-  font-size: 30px;
-  background-color: #9f9f9f;
-  color: #fff;
-  border-radius: var(--lay-border-radius);
-  opacity: 0.95;
-}
-.layui-fixbar li:hover {
-  opacity: 0.85;
-}
-.layui-fixbar li:active {
-  opacity: 1;
-}
-.layui-fixbar .layui-fixbar-top {
-  display: none;
-  font-size: 40px;
-}
