+++ src/index.js
@@ -11,10 +11,13 @@ export { router } from './core/router.js';
 export { default as jquery, default as $ } from 'jquery';
 export { component, componentBuilder } from './core/component.js';
 
+// 导出工具模块
+export * as utils from './utils/index.js';
+
 // 导出组件模块
-export { laypage } from './components/laypage.js';
-export { laydate } from './components/laydate.js';
 export { layer } from './components/layer.js';
+export { laydate } from './components/laydate.js';
+export { laypage } from './components/laypage.js';
 export { dropdown } from './components/dropdown.js';
 export { slider } from './components/slider.js';
 export { colorpicker } from './components/colorpicker.js';
@@ -33,6 +36,5 @@ export { transfer } from './components/transfer.js';
 export { carousel } from './components/carousel.js';
 export { rate } from './components/rate.js';
 export { flow } from './components/flow.js';
-export { util } from './components/util.js';
 export { floatbar } from './components/floatbar.js';
 export { code } from './components/code.js';
