+++ src/utils/index.js
@@ -1,0 +1,8 @@
+/**
+ * utils
+ * 工具集
+ */
+
+export * from '../utils/events.js';
+export * from '../utils/date.js';
+export * from '../utils/browser.js';
