+++ tests/visual/assets/i18n/zh-HK.js
@@ -80,7 +80,7 @@ export default {
     cancel: '取消',
     defaultTitle: '資訊',
     prompt: {
-      InputLengthPrompt: '最多輸入 {length} 個字符',
+      inputLengthPrompt: '最多輸入 {length} 個字符',
     },
     photos: {
       noData: '沒有圖片',
@@ -161,7 +161,7 @@ export default {
     },
     chooseText: '{length} 個檔案',
   },
-  util: {
+  utils: {
     timeAgo: {
       days: '{days} 天前',
       hours: '{hours} 小時前',
