+++ tests/visual/tabs.html
@@ -217,7 +217,7 @@
       layui.use(() => {
         var $ = layui.$;
         var tabs = layui.tabs;
-        var util = layui.util;
+        var utils = layui.utils;
         var layer = layui.layer;
         var dropdown = layui.dropdown;
 
@@ -251,7 +251,7 @@
         };
 
         // 自定义事件
-        util.on({
+        utils.on({
           add: () => {
             addTabs();
           },
