+++ tests/visual/template/release/test.html
@@ -39,7 +39,7 @@
     <script src="./dist/layui.js"></script>
     <script type="module">
       layui.use(() => {
-        const { lay, layer, form, laydate, util } = layui;
+        const { lay, layer, form, laydate, utils } = layui;
 
         // 欢迎信息
         layer.msg('Hello World', { icon: 6 });
@@ -55,7 +55,7 @@
         });
 
         // 触发事件
-        util.on({
+        utils.on({
           'test-form': function () {
             layer.open({
               type: 1,
@@ -63,28 +63,28 @@
               shadeClose: true,
               area: '350px',
               title: 'layer + form',
-              content: [
-                '<ul class="layui-form layui-form-pane" style="margin: 15px;">',
-                '<li class="layui-form-item">',
-                '<label class="layui-form-label">输入框</label>',
-                '<div class="layui-input-block">',
-                '<input class="layui-input" lay-verify="required" name="field1">',
-                '</div>',
-                '</li>',
-                '<li class="layui-form-item">',
-                '<label class="layui-form-label">选择框</label>',
-                '<div class="layui-input-block">',
-                '<select name="field2">',
-                '<option value="A">A</option>',
-                '<option value="B">B</option>',
-                '</select>',
-                '</div>',
-                '</li>',
-                '<li class="layui-form-item" style="text-align:center;">',
-                '<button type="submit" lay-submit lay-filter="*" class="layui-btn">提交</button>',
-                '</li>',
-                '</ul>',
-              ].join(''),
+              content: `
+<ul class="layui-form layui-form-pane" style="margin: 15px;">
+  <li class="layui-form-item">
+    <label class="layui-form-label">输入框</label>
+    <div class="layui-input-block">
+      <input class="layui-input" lay-verify="required" name="field1">
+    </div>
+  </li>
+  <li class="layui-form-item">
+    <label class="layui-form-label">选择框</label>
+    <div class="layui-input-block">
+      <select name="field2">
+        <option value="A">A</option>
+        <option value="B">B</option>
+      </select>
+    </div>
+  </li>
+  <li class="layui-form-item" style="text-align:center;">
+    <button type="submit" lay-submit lay-filter="*" class="layui-btn">提交</button>
+  </li>
+</ul>
+              `.trim(),
               success: function (layero, index) {
                 layero.find('.layui-layer-content').css('overflow', 'visible');
 
@@ -92,7 +92,7 @@
                   var field = data.field;
 
                   // 显示填写的表单
-                  layer.msg(util.escape(JSON.stringify(field)), {
+                  layer.msg(lay.escape(JSON.stringify(field)), {
                     icon: 1,
                   });
                   // layer.close(index); //关闭层
